Fix Windows build errors and deprecation warnings
- Fix type mismatches in Windows memory monitoring (cast usize to u64)
- Replace deprecated child_ui with new_child using UiBuilder
- Fix unused variable warning by prefixing with underscore

The code now builds cleanly on both Linux and Windows platforms.

+++ src/memory_monitor.rs
@@ -243,8 +243,8 @@ fn get_memory_info() -> Option<MemoryInfo> {
         }
         
         Some(MemoryInfo {
-            physical_memory: pmc.WorkingSetSize,
-            virtual_memory: pmc.PagefileUsage,
+            physical_memory: pmc.WorkingSetSize as u64,
+            virtual_memory: pmc.PagefileUsage as u64,
             system_total: mem_status.ullTotalPhys,
             system_available: mem_status.ullAvailPhys,
         })
